feat: migrate fallback.test.js to jest
